+++ spigot-utils/src/main/java/com/kamikazejam/kamicommon/SpigotUtilsSource.java
@@ -4,8 +4,8 @@ import com.kamikazejam.kamicommon.command.KamiCommand;
 import com.kamikazejam.kamicommon.command.KamiCommonCommandRegistration;
 import com.kamikazejam.kamicommon.command.impl.kc.KamiCommonCommand;
 import com.kamikazejam.kamicommon.configuration.spigot.KamiConfig;
-import com.kamikazejam.kamicommon.gui.MenuManager;
-import com.kamikazejam.kamicommon.gui.MenuTask;
+import com.kamikazejam.kamicommon.menu.MenuManager;
+import com.kamikazejam.kamicommon.menu.MenuTask;
 import com.kamikazejam.kamicommon.integrations.PlaceholderAPIIntegration;
 import com.kamikazejam.kamicommon.integrations.PremiumVanishIntegration;
 import com.kamikazejam.kamicommon.nms.NmsVersion;
