--- spigot-utils/src/main/java/com/kamikazejam/kamicommon/gui/clicks/MenuClickEvent.java
@@ -1,11 +1,0 @@
-package com.kamikazejam.kamicommon.gui.clicks;
-
-import org.bukkit.entity.Player;
-import org.bukkit.event.inventory.ClickType;
-import org.bukkit.event.inventory.InventoryClickEvent;
-import org.jetbrains.annotations.NotNull;
-
-@SuppressWarnings("unused")
-public interface MenuClickEvent {
-    void onClick(@NotNull Player player, @NotNull ClickType clickType, @NotNull InventoryClickEvent event);
-}
