--- spigot-utils/src/main/java/com/kamikazejam/kamicommon/gui/clicks/transform/MenuClickTransform.java
@@ -1,22 +1,0 @@
-package com.kamikazejam.kamicommon.gui.clicks.transform;
-
-import com.kamikazejam.kamicommon.gui.clicks.MenuClick;
-import org.bukkit.entity.Player;
-import org.bukkit.event.inventory.InventoryClickEvent;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.Objects;
-
-@SuppressWarnings("unused")
-public class MenuClickTransform implements IClickTransform {
-
-    private final @NotNull MenuClick click;
-    public MenuClickTransform(@NotNull MenuClick click) {
-        this.click = Objects.requireNonNull(click);
-    }
-
-    @Override
-    public void process(@NotNull Player player, @NotNull InventoryClickEvent event, int page) {
-        click.onClick(player, event.getClick());
-    }
-}
