--- spigot-utils/src/main/java/com/kamikazejam/kamicommon/gui/items/interfaces/IBuilderModifier.java
@@ -1,8 +1,0 @@
-package com.kamikazejam.kamicommon.gui.items.interfaces;
-
-import com.kamikazejam.kamicommon.item.IBuilder;
-import org.jetbrains.annotations.NotNull;
-
-public interface IBuilderModifier {
-    void modify(@NotNull IBuilder builder);
-}
