--- spigot-utils/src/main/java/com/kamikazejam/kamicommon/gui/items/slots/ItemSlot.java
@@ -1,11 +1,0 @@
-package com.kamikazejam.kamicommon.gui.items.slots;
-
-import com.kamikazejam.kamicommon.gui.KamiMenu;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.Set;
-
-public interface ItemSlot {
-    Set<Integer> get(@NotNull KamiMenu menu);
-    @NotNull ItemSlot copy();
-}
