+++ spigot-utils/src/main/java/com/kamikazejam/kamicommon/menu/impl/SimpleMenu.java
@@ -1,0 +1,9 @@
+package com.kamikazejam.kamicommon.menu.impl;
+
+/**
+ * This Menu class focuses on providing a simple single-frame menu. This is the most versatile menu type
+ * because you define everything in the menu, and can create your own custom logic.
+ */
+public class SimpleMenu {
+
+}
