+++ spigot-utils/src/main/java/com/kamikazejam/kamicommon/menu/items/interfaces/IBuilderModifier.java
@@ -1,0 +1,8 @@
+package com.kamikazejam.kamicommon.menu.items.interfaces;
+
+import com.kamikazejam.kamicommon.item.IBuilder;
+import org.jetbrains.annotations.NotNull;
+
+public interface IBuilderModifier {
+    void modify(@NotNull IBuilder builder);
+}
